Add Spanish and German translations for i18n
The code changes in this commit add Spanish and German translations to the i18n module. The `es.json` and `de.json` files are imported and included in the resources object of the i18n configuration. This enhancement allows the application to support multiple languages, improving the user experience for Spanish and German-speaking users.